requirements.txt
rembg==0.0.58
Pillow>=10.0.0
tqdm>=4.66.0
